Move dmarc management to its own file
