+++ app/errors.py
@@ -56,8 +56,3 @@ class MailSentFromReverseAlias(SLException):
     """raised when receiving an email sent from a reverse alias"""
 
     pass
-
-
-class DmarcSoftFail(SLException):
-
-    pass
