+++ tests/test_email_handler.py
@@ -104,25 +104,17 @@ def test_dmarc_forward_quarantine(flask_client):
     assert f"{alias.email} has a new mail in quarantine" == notifications[0].title
 
 
-# todo: re-enable test when softfail is quarantined
-# def test_gmail_dmarc_softfail(flask_client):
-#     user = create_random_user()
-#     alias = Alias.create_new_random(user)
-#     msg = load_eml_file("dmarc_gmail_softfail.eml", {"alias_email": alias.email})
-#     envelope = Envelope()
-#     envelope.mail_from = msg["from"]
-#     envelope.rcpt_tos = [msg["to"]]
-#     result = email_handler.handle(envelope, msg)
-#     assert result == status.E215
-#     email_logs = (
-#         EmailLog.filter_by(user_id=user.id, alias_id=alias.id)
-#         .order_by(EmailLog.id.desc())
-#         .all()
-#     )
-#     assert len(email_logs) == 1
-#     email_log = email_logs[0]
-#     assert email_log.blocked
-#     assert email_log.refused_email_id
+def test_gmail_dmarc_softfail(flask_client):
+    user = create_random_user()
+    alias = Alias.create_new_random(user)
+    msg = load_eml_file("dmarc_gmail_softfail.eml", {"alias_email": alias.email})
+    envelope = Envelope()
+    envelope.mail_from = msg["from"]
+    envelope.rcpt_tos = [msg["to"]]
+    result = email_handler.handle(envelope, msg)
+    assert result == status.E200
+    payload = msg.get_payload()
+    assert payload.find("failed anti-phishing checks") > -1
 
 
 def test_prevent_5xx_from_spf(flask_client):
